+++ lib/LatteExtension.php
@@ -3,6 +3,7 @@
 namespace JanHerman\Barista;
 
 use Latte\Extension;
+use Latte\Runtime\FilterInfo;
 
 class LatteExtension extends Extension
 {
@@ -13,7 +14,17 @@ class LatteExtension extends Extension
 
     public function getFilters(): array
     {
-        return option('jan-herman.barista.filters', []);
+        $built_in_filters = [
+            'stripNewLines' => function (FilterInfo $info, string $html): string {
+                $lines = preg_split('/\R/', $html);
+                $lines = array_map('ltrim', $lines);
+                return implode('', $lines);
+            }
+        ];
+
+        $custom_filters = option('jan-herman.barista.filters', []);
+
+        return array_merge($built_in_filters, $custom_filters);
     }
 
     public function getFunctions(): array